requirements.txt
python-dateutil
python-dotenv
prettyTable
flask
flask-sqlalchemy
gunicorn
mysqlclient
pandas
